some warnings and messages
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@131 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfrelay/bzfrelay.c
@@ -1029,7 +1029,7 @@ static int		daemonize(void)
     rl.rlim_max = 1024;
 
   /* close all files except stderr */
-  for (i = 0; i < rl.rlim_max; i++)
+  for (i = 0; i < (int)rl.rlim_max; i++)
     if (i != 2)
       if (close(i) < 0 && errno != EBADF)
 	return -1;
